k

Struct JacobianIkSolver

pub struct JacobianIkSolver<T: RealField> {
    pub allowable_target_distance: T,
    pub allowable_target_angle: T,
    pub jacobian_multiplier: T,
    pub num_max_try: usize,
    /* private fields */
}

Inverse Kinematics Solver using Jacobian matrix

Fields
allowable_target_distance: T

If the distance is smaller than this value, it is reached.

allowable_target_angle: T

If the angle distance is smaller than this value, it is reached.

jacobian_multiplier: T

multiplier for jacobian

num_max_try: usize

How many times the joints are tried to be moved

Implementations

impl<T> JacobianIkSolver<T>
where T: RealField + SubsetOf<f64>,

pub fn new( allowable_target_distance: T, allowable_target_angle: T, jacobian_multiplier: T, num_max_try: usize, ) -> JacobianIkSolver<T>

Create instance of JacobianIkSolver.

JacobianIkSolverBuilder is available instead of calling this new method.

Examples

let solver = k::JacobianIkSolver::new(0.01, 0.01, 0.5, 100);

pub fn set_nullspace_function( &mut self, func: Box<dyn Fn(&[T]) -> Vec<T> + Send + Sync>, )

Set a null space function for redundant manipulator.

Examples

let mut solver = k::JacobianIkSolver::new(0.01, 0.01, 0.5, 100);
solver.set_nullspace_function(Box::new(
k::create_reference_positions_nullspace_function(
   vec![0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0],
   vec![0.1, 0.1, 0.1, 1.0, 0.1, 0.5, 0.0],
   ),
));

pub fn clear_nullspace_function(&mut self)

Clear the null function which is set by set_nullspace_function.