k

Struct SerialChain

pub struct SerialChain<T: RealField> { /* private fields */ }

Kinematic chain without any branch.

All joints are connected sequentially.

Implementations

impl<T> SerialChain<T>
where T: RealField + SubsetOf<f64>,

pub fn new_unchecked(inner: Chain<T>) -> Self

Convert Chain into SerialChain without any check

If the input Chain has any branches it causes serious bugs.

pub fn try_new(inner: Chain<T>) -> Option<Self>

Convert Chain into SerialChain

If the input Chain has any branches it fails.

Examples

let node = k::NodeBuilder::<f32>::new().into_node();
let chain = k::Chain::from_root(node);
assert!(k::SerialChain::try_new(chain).is_some());

let node0 = k::NodeBuilder::<f32>::new().into_node();
let node1 = k::NodeBuilder::new().into_node();
let node2 = k::NodeBuilder::new().into_node();
node1.set_parent(&node0);
node2.set_parent(&node0);
let chain = k::Chain::from_root(node0);
assert!(k::SerialChain::try_new(chain).is_none());

pub fn from_end(end_joint: &Node<T>) -> SerialChain<T>

Create SerialChain from the end Node

Examples

let node = k::NodeBuilder::<f32>::new().into_node();
let s_chain = k::SerialChain::from_end(&node);

pub fn from_end_to_root( end_joint: &Node<T>, root_joint: &Node<T>, ) -> SerialChain<T>

Create SerialChain from the end Node and root Node.

Examples

let node0 = k::NodeBuilder::<f32>::new().into_node();
let node1 = k::NodeBuilder::<f32>::new().into_node();
let node2 = k::NodeBuilder::<f32>::new().into_node();
let node3 = k::NodeBuilder::<f32>::new().into_node();
use k::connect;
connect![node0 => node1 => node2 => node3];
let s_chain = k::SerialChain::from_end_to_root(&node2, &node1);
assert_eq!(s_chain.iter().count(), 2);

pub fn unwrap(self) -> Chain<T>

Safely unwrap and returns inner Chain instance

pub fn end_transform(&self) -> Isometry3<T>

Calculate transform of the end joint