k

Trait InverseKinematicsSolver

pub trait InverseKinematicsSolver<T>
where
    T: RealField,
{
    // Required method
    fn solve_with_constraints(
        &self,
        arm: &SerialChain<T>,
        target_pose: &Isometry3<T>,
        constraints: &Constraints,
    ) -> Result<(), Error>;

    // Provided method
    fn solve(
        &self,
        arm: &SerialChain<T>,
        target_pose: &Isometry3<T>,
    ) -> Result<(), Error> { ... }
}

IK solver

Required Methods

fn solve_with_constraints( &self, arm: &SerialChain<T>, target_pose: &Isometry3<T>, constraints: &Constraints, ) -> Result<(), Error>

Move the end transform of the arm to target_pose with constraints

Provided Methods

fn solve( &self, arm: &SerialChain<T>, target_pose: &Isometry3<T>, ) -> Result<(), Error>

Move the end transform of the arm to target_pose